test: split design component tests and align fixtures Related to: #112
